# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "VTuberland - API SPECIFICATIONS",
        "description": "An API documentation used for VTuberland App",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "https://vtuberland.test/api/v1/",
            "description": "Local Server"
        },
        {
            "url": "http://vtuberland.commude.biz/api/v1/",
            "description": "Dev Server"
        }
    ],
    "paths": {
        "/auth/login": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Login User",
                "description": "User Login",
                "operationId": "App\\Http\\Controllers\\User\\Api\\AuthController::login",
                "parameters": [
                    {
                        "name": "email",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "password",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Auth"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                },
                "security": [
                    {
                        "passport": [
                            "*"
                        ]
                    }
                ]
            }
        },
...
        "/me": {
            "get": {
                "tags": [
                    "Me"
                ],
                "summary": "Get current user",
                "description": "Get current details of the user.",
                "operationId": "App\\Http\\Controllers\\User\\Api\\MeController::index",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/User"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                },
                "security": [
                    {
                        "passport": [
                            "*"
                        ]
                    }
                ]
            },
            "post": {
                "tags": [
                    "Me"
                ],
                "summary": "Register User",
                "description": "Create a new user.",
                "operationId": "App\\Http\\Controllers\\User\\Api\\MeController::register",
                "parameters": [
                    {
                        "name": "name",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "email",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "password",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "password_confirmation",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "manufacturer",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "os",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "version",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "language",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "device_uuid",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Auth"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                },
                "security": [
                    {
                        "passport": [
                            "*"
                        ]
                    }
                ]
            }
        },
        "/me/logout": {
            "post": {
                "tags": [
                    "Me"
                ],
                "summary": "Logout User",
                "description": "Revoke the current session of user.",
                "operationId": "App\\Http\\Controllers\\User\\Api\\MeController::logout",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "code": {
                                            "description": "Revoke active token of the current user.",
                                            "type": "integer",
                                            "format": "integer",
                                            "example": 200
                                        },
                                        "message": {
                                            "description": "Revoke active token of the current user.",
                                            "type": "text",
                                            "format": "string",
                                            "example": "ユーザーは正常にログアウトしました"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                },
                "security": [
                    {
                        "passport": [
                            "*"
                        ]
                    }
                ]
            }
        },
        "/me/refresh": {
            "post": {
                "tags": [
                    "Me"
                ],
                "summary": "Refresh user session.",
                "description": "Revoke and Generate new user session.",
                "operationId": "App\\Http\\Controllers\\User\\Api\\MeController::refresh",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Auth"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                },
                "security": [
                    {
                        "passport": [
                            "*"
                        ]
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Auth": {
                "properties": {
                    "access_token": {
                        "description": "Transform the resource into an array.",
                        "type": "string",
                        "format": "string",
                        "example": "eyJ0eXAiOiJKV1QiLCJhbGciOiJSUzI1NiJ9.eyJhdWQiOiI5MTc4ZDY4NC0yYWFkLTRkNjItODQ0Yi0zZTNlZDJjZDZlOTIiLCJqdGkiOiI2NDgyY2YzZjQ5MjA3OTA1YjM2ZDc2YWZlOGM5YWRmYTk0M2MwMzBlNjRmMGU3YjhjZjcwMDNmNWUwM2MyMDYxZjJkNTc4NTMyOGI2NDA5MyIsImlhdCI6MTU5OTQ4Mzg4OCwibmJmIjoxNTk5NDgzODg4LCJleHAiOjE2MzEwMTk4ODgsInN1YiI6ImJkY2E5MzQ4LWNjZjItNDliNy05NzQzLWQxNzM0ODg3NmZlOSIsInNjb3BlcyI6W119.N8owVNbUQNpUREfRHoalOV-FLVbGFURBZ_Ez9eyaiItM8tg1qJZ5PhRrGIaUvTyhEHzoqoa9EMhqvfamrbpghx7tAJscS-Xd1CM2nh0g-3LetKXHvvwbYGQ3D8nw5hlfwLlSHfx-YlH6WN4sDjT36rdbc5IG0Qxdn7GN8JQu3bJWDguzcbiuyiph-VS-0smankid4h7oo1siFEuS96X-7E2-SN6qm4s0v7M_IEb9rL9iMTp0mdU4KG9ZFj31C6wdGo0uotT7osdQ43nEJak6cxNNys8NMhn46YqQw2VFG4fYdXYbQ8o7Gqj8aqDRo5oLKcVBXc4fPRVq6tbnVkogUoZxDCfVMS6AWYthbEnd2Y7R12vP53ZPQ1MDNM-n3Px8syaj8ovx3thfZy6dsdH914qCHqo3nfyxWZo6TXegDqeAGhAB1cDnU-fk6NwEmp1HzKVEJaoW8MxDIMVYG8cbd0SViT7pX6LVJUPe9LIDEVaTdLB_Oek2eXht51DXv26WWvaZpRbZd9aLmoM5krceBFvy-z8WXwRVaSKOFBxZgNR5Avnp2a5CJVvDGKFZ2r_8otr4WD04qxBIc16ystzJMeDtFyyZRpaSHMl2ehhaEAbAXvzU7jCmvzHbUzAo_8W_BJiI4IZkSrwrgSQUWVgP2W1kS1fA8jul37-opNWOeIE"
                    },
                    "token_type": {
                        "description": "Transform the resource into an array.",
                        "type": "string",
                        "format": "string",
                        "example": "Bearer"
                    },
                    "expires_at": {
                        "description": "Transform the resource into an array.",
                        "type": "timestamp",
                        "format": "date",
                        "example": "2020-03-10T19:42:31+09:00"
                    }
                },
                "type": "object"
            },
            "User": {
                "properties": {
                    "id": {
                        "description": "Transform the resource into an array.",
                        "type": "uuid",
                        "format": "string",
                        "example": "bdca9348-ccf2-49b7-9743-d17348876fe9"
                    },
                    "name": {
                        "description": "Transform the resource into an array.",
                        "type": "string",
                        "format": "string",
                        "example": "Takemoto Iwao"
                    },
                    "email": {
                        "description": "Transform the resource into an array.",
                        "type": "email",
                        "format": "string",
                        "example": "ultimate000@example.com"
                    },
                    "manufacturer": {
                        "description": "Transform the resource into an array.",
                        "type": "string",
                        "format": "string",
                        "example": "apple"
                    },
                    "os": {
                        "description": "Transform the resource into an array.",
                        "type": "string",
                        "format": "string",
                        "example": "iOS"
                    },
                    "version": {
                        "description": "Transform the resource into an array.",
                        "type": "string",
                        "format": "string",
                        "example": "1.0.0"
                    },
                    "language": {
                        "description": "Transform the resource into an array.",
                        "type": "string",
                        "format": "string",
                        "example": "ja_JP"
                    },
                    "device_uuid": {
                        "description": "Transform the resource into an array.",
                        "type": "string",
                        "format": "string",
                        "example": "26397745-9126-438c-9706-5002baf5d3a4"
                    },
                    "is_valid": {
                        "description": "Transform the resource into an array.",
                        "type": "boolean",
                        "format": "boolean",
                        "example": 1
                    },
                    "created_at": {
                        "description": "Transform the resource into an array.",
                        "type": "timestamp",
                        "format": "date",
                        "example": "2020-03-10T19:42:31+09:00"
                    }
                },
                "type": "object"
            },
...
        }
    },
    "security": [
        []
    ]
}